Address comments #2
